Checked whether subterms can be easily picked from the tree.
We need to be able to pick subterms in order to apply substitution to the trees. In order not to get to a node by accessing its child, and its child child and its child child child, I'll implement getting a node by its position on the tree. Substitution can be made then by picking the subterm starting from the node at position X.

Signed-off-by: Maria-Minerva Vonica <maria.vonica98@gmail.com>